Add migration only if needed and add grade to food
Migration only happens if the database does not have any food categories
in the table. Note that this does not check whether the rows are correct
compared to the csv.

Add a simple grade to food. The nutrients required each day varies,
depending on sex and whether they are pregnant or lactating. An "Other"
option is selected if the person enters an invalid state. Data is
provided by NIH and FDA.